pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Projeto repositório de funcionalidades."
authors = [
    {name = "djangonauta", email = "igbrch@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "arrow (>=1.3.0,<2.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "crispy-bootstrap5 (>=2025.6,<2026.0)",
    "django (>=5.2.6,<6.0.0)",
    "django-allauth (>=65.11.2,<66.0.0)",
    "django-auditlog (>=3.2.1,<4.0.0)",
    "django-auth-ldap (>=5.2.0,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-cors-headers (>=4.8.0,<5.0.0)",
    "django-crispy-forms (>=2.4,<3.0)",
    "django-csp (>=4.0,<5.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-extra-fields (>=3.0.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-formtools (>=2.5.1,<3.0.0)",
    "django-guardian (>=3.2.0,<4.0.0)",
    "django-hijack (>=3.7.4,<4.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "django-pipeline (>=4.1.0,<5.0.0)",
    "django-post-office (>=3.10.1,<4.0.0)",
    "django-prometheus (>=2.4.1,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "django-rest-auth (>=0.9.5,<0.10.0)",
    "django-select2 (>=8.4.1,<9.0.0)",
    "django-structlog[celery] (>=9.1.1,<10.0.0)",
    "django-view-breadcrumbs (>=2.5.1,<3.0.0)",
    "django-weasyprint (>=2.4.0,<3.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-guardian2 (>=0.7.0,<0.8.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.9.1,<2026.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "hiredis (>=3.2.1,<4.0.0)",
    "invoke (>=2.2.0,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "pygraphviz (>=1.14,<2.0)",
    "redis (>=6.4.0,<7.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "weasyprint (>=66.0,<67.0)",
    "whitenoise (>=6.10.0,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
commitizen = "^4.9.1"
coverage = "^7.10.6"
django-debug-toolbar = "^6.0.0"
fabric = "^3.2.2"
factory-boy = "^3.3.3"
flake8 = "^7.3.0"
ipython = "^9.5.0"
isort = "^6.0.1"
pyopenssl = "^25.2.0"
selenium = "^4.35.0"
werkzeug = "^3.1.3"

[tool.black]
line-length = 100
extend-exclude = '''
/(
  migrations
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver2"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
version_files = ["pyproject.toml:project.version"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
